Fix an issue where the args list would get corrupted if Procs were used.
The issue seems to center around the union types that mruby uses.
Examining the anomaly I noticed the proc values were getting truncated
-- they are a part of a union and maybe mruby is doing something weird
here.

Either way, copying the struct (leaving the data unaltered) as opposed
to passing the pointer, seems to preserve the width information which is
what was causign the issue. I strongly believe this fixes two bugs; one
in argument handling of proc and another where certain values would get
corrupted and could not be extracted as string.

The rest of this is cleanup and rearch for simpler code.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>