Fix exception propagation when N > 1 calls are calling in/out of go-mruby and one of then throws.
The offender looks something like this (but in go callbacks):

def test args
  raise "test"
end

def test2 args
  test args
end

If both of these are go-backed functions (just imagine) and `test`
raises during the invocation of `test2`, we would set the exception in
the second run through, raise it, then clear it in the first, causing
mruby to get really confused trying to handle an exception that no
longer exists.

The code we were using to do this was really dodgy WRT struct/union
cooperation so I reduced it to behave the same way like callouts do, and
wrote some code to massage the exception back into the state. It seems
to work fine.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>